@@ -13,7 +13,7 @@ authors = [
 ]
 description = "A Python client for RTMA/Dragonfly"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
