@@ -1,97 +1,124 @@
 import time
-import json
+import threading
 import pyrtma
+import pathlib
 import pyrtma.core_defs as cd
+
+from pyrtma.data_logger.dataset import Dataset
 
 
 def main(trial_num: int, count: int):
+    # Reset the DataLogger for a clean start
     mod = pyrtma.Client()
     mod.connect()
-
     mod.send_signal(cd.MT_DATA_LOGGER_RESET)
 
-    # Add a DataCollection with a DataSet
-    msg = cd.MDF_ADD_DATA_COLLECTION()
-    msg.collection.name = "test"
-    msg.collection.base_path = f"$(env:USERPROFILE)\\Desktop"
-    msg.collection.dir_fmt = "test_$(date)"
-    msg.collection.num_data_sets = 1
-    msg.collection.data_sets[0].name = "QL.Data"
-    msg.collection.data_sets[0].sub_dir_fmt = "trial_$(trial_num:04d)/QL"
-    msg.collection.data_sets[0].file_name_fmt = f"QL.Data_trial_$(trial_num:04d)"
-    msg.collection.data_sets[0].formatter = "quicklogger"
-    msg.collection.data_sets[0].msg_types[0] = cd.ALL_MESSAGE_TYPES
-    msg.collection.data_sets[0].subdivide_interval = 30
-    mod.send_message(msg)
+    # Add Datasets to the DataLogger
+    base_path = pathlib.Path.home() / "data_logger"
 
-    # Add/Update DataSets to an existing DataCollection
+    datasets: list[Dataset] = list()
 
     # JSONFormatter
-    msg = cd.MDF_ADD_DATA_SET()
-    msg.collection_name = "test"
-    msg.data_set.name = "json"
-    msg.data_set.sub_dir_fmt = "trial_$(trial_num:04d)/json"
-    msg.data_set.file_name_fmt = f"json_trial_$(trial_num:04d)"
-    msg.data_set.formatter = "json"
-    msg.data_set.msg_types[0] = cd.MT_DATA_LOG_TEST_2048
-    msg.data_set.subdivide_interval = 0
-    mod.send_message(msg)
+    jsonf_ds = Dataset(
+        name="json",
+        save_path=base_path / f"trial_{trial_num:04d}" / "json",
+        filename=f"trial_{trial_num:04d}",
+        formatter="json",
+        msg_types=[cd.MT_DATA_LOG_TEST_2048],
+        subdivide_interval=30,
+    )
+
+    # Could also reset before adding a Dataset
+    jsonf_ds.reset_data_logger()
+
+    jsonf_ds.add()
+    datasets.append(jsonf_ds)
 
     # MsgHeaderFormatter
-    msg = cd.MDF_ADD_DATA_SET()
-    msg.collection_name = "test"
-    msg.data_set.name = "headers"
-    msg.data_set.sub_dir_fmt = "trial_$(trial_num:04d)/msg_headers"
-    msg.data_set.file_name_fmt = f"msg_headers_trial_$(trial_num:04d)"
-    msg.data_set.formatter = "msg_header"
-    msg.data_set.msg_types[0] = cd.MT_DATA_LOG_TEST_2048
-    msg.data_set.subdivide_interval = 0
-    mod.send_message(msg)
+    headersf_ds = Dataset(
+        name="headers",
+        save_path=str(base_path / f"trial_{trial_num:04d}" / "headers"),
+        filename=f"trial_{trial_num:04d}",
+        formatter="msg_header",
+        msg_types=[cd.MT_DATA_LOG_TEST_2048],
+        subdivide_interval=0,
+    )
+    headersf_ds.add()
+    datasets.append(headersf_ds)
 
-    # RawFormatter
-    msg = cd.MDF_ADD_DATA_SET()
-    msg.collection_name = "test"
-    msg.data_set.name = "raw"
-    msg.data_set.sub_dir_fmt = "trial_$(trial_num:04d)/raw"
-    msg.data_set.file_name_fmt = f"raw_trial_$(trial_num:04d)"
-    msg.data_set.formatter = "raw"
-    msg.data_set.msg_types[0] = cd.MT_DATA_LOG_TEST_2048
-    mod.send_message(msg)
+    # QLFormatter
+    qlf_ds = Dataset(
+        name="quicklogger",
+        save_path=str(base_path / f"trial_{trial_num:04d}" / "quicklogger"),
+        filename=f"trial_{trial_num:04d}",
+        formatter="quicklogger",
+        msg_types=[cd.MT_DATA_LOG_TEST_2048],
+    )
+    qlf_ds.add()
+    datasets.append(qlf_ds)
 
-    # Send the metadata for trial_num
-    msg = cd.MDF_DATA_LOGGER_METADATA_UPDATE()
-    meta = dict(trial_num=trial_num)
-    msg.json = json.dumps(meta, separators=(",", ":"))
-    mod.send_message(msg)
+    # RawFormatter (Alternative binary format [Header->Data->Header->Data])
+    # rawf_ds = Dataset(
+    #     name="raw",
+    #     save_path=str(base_path / f"trial_{trial_num:04d}" / "raw"),
+    #     filename=f"trial_{trial_num:04d}",
+    #     formatter="raw",
+    #     msg_types=[cd.MT_DATA_LOG_TEST_2048]
+    # )
+    # datasets.append(raw_ds)
 
-    mod.send_signal(cd.MT_DATA_LOGGER_START)
+    # Start all the data sets
+    input("Hit enter to start the data sets...")
+    for ds in datasets:
+        ds.start()
+        print(str(ds))
+
     input("Hit enter to start sending packets...")
-    msg = cd.MDF_DATA_LOG_TEST_2048.from_random()
-    n = 0
+    stop = threading.Event()
+
+    def thread_func(count):
+        msg = cd.MDF_DATA_LOG_TEST_2048.from_random()
+        n = 0
+        try:
+            if count > 0:
+                print(f"Sending {count} packets...")
+                for _ in range(count):
+                    mod.send_message(msg)
+                    n += 1
+
+                    tic = time.perf_counter()
+                    while (time.perf_counter() - tic) < 0.002:
+                        pass
+            else:
+                while not stop.is_set():
+                    mod.send_message(msg)
+                    n += 1
+
+                    tic = time.perf_counter()
+                    while (time.perf_counter() - tic) < 0.002:
+                        pass
+        finally:
+            print(f"Sent {n} packets.")
+            stop.set()
+
+    t = threading.Thread(target=thread_func, args=(count,))
+    t.start()
+
+    # Monitor our data collection
     try:
-        if count > 0:
-            print(f"Sending {count} packets...")
-            for _ in range(count):
-                mod.send_message(msg)
-                n += 1
+        while not stop.is_set():
+            for ds in datasets:
+                ds.poll()
 
-                tic = time.perf_counter()
-                while (time.perf_counter() - tic) < 0.002:
-                    pass
-        else:
-            while True:
-                mod.send_message(msg)
-                n += 1
-
-                tic = time.perf_counter()
-                while (time.perf_counter() - tic) < 0.002:
-                    pass
+            time.sleep(0.250)
     except KeyboardInterrupt:
-        pass
+        stop.set()
     finally:
-        print(f"Sent {n} packets.")
-        if mod.connected:
-            mod.send_signal(cd.MT_DATA_LOGGER_STOP)
+        for ds in datasets:
+            ds.stop()
+            print(str(ds))
+
+        t.join()
 
     print("Done")
 
