@@ -79,21 +79,8 @@ class MatlabDefCompiler:
     def generate_fcn_close() -> str:
         return "end\n"
 
-    @staticmethod
-    def sanitize_name(name: str) -> str:
-        # strip leading characters that are invalid to start
-        #  a fieldname (only letters allowed)
-        name = name.lstrip("_0123456789")
-
-        # matlab fieldnames allow alphanum and _ after starting characters
-        for c in name:
-            if not c.isalnum() and c != "_":
-                name = name.replace(c, "")
-        return name
-
     def generate_field(self, top_field: str, name: str, value: Any) -> str:
         name = name.replace(f"{top_field}_", "", 1)  # strip top_field from fieldname
-        name = self.sanitize_name(name)
         return f"{self.struct_name}.{top_field}.{name} = {value};\n"
 
     def generate_constant(self, c: Union[ConstantExpr, HID, MID, MT]) -> str:
@@ -105,25 +92,23 @@ class MatlabDefCompiler:
             prefix = "MID_"
         elif isinstance(c, MT):
             prefix = "MT_"
-        return self.generate_field(
-            "defines", f"{prefix}{self.sanitize_name(c.name)}", c.value
-        )
+        return self.generate_field("defines", f"{prefix}{c.name}", c.value)
 
     def generate_constant_string(self, c: ConstantString):
-        return self.generate_field("defines", self.sanitize_name(c.name), c.value)
+        return self.generate_field("defines", c.name, c.value)
 
     def generate_host_id(self, hid: HID) -> str:
-        return self.generate_field("HID", self.sanitize_name(hid.name), hid.value)
+        return self.generate_field("HID", hid.name, hid.value)
 
     def generate_module_id(self, mid: MID) -> str:
-        return self.generate_field("MID", self.sanitize_name(mid.name), mid.value)
+        return self.generate_field("MID", mid.name, mid.value)
 
     def generate_msg_type_id(self, mt: MT) -> str:
-        return self.generate_field("MT", self.sanitize_name(mt.name), mt.value)
+        return self.generate_field("MT", mt.name, mt.value)
 
     def generate_type_alias(self, td: TypeAlias) -> str:
-        name = self.sanitize_name(td.name)
-        type_name = self.sanitize_name(td.type_name)
+        name = td.name
+        type_name = td.type_name
         if td.type_name in type_map.keys():
             s = type_map[td.type_name]
             return f"{self.struct_name}.typedefs.{name} = {s}(0);\n"
@@ -144,7 +129,7 @@ class MatlabDefCompiler:
     ) -> str:
         f = []
 
-        name = self.sanitize_name(struct.name)
+        name = struct.name
 
         if isinstance(struct, MDF) and len(struct.fields) == 0:
             f.append(f"% {name} (Signal)")
@@ -184,7 +169,7 @@ class MatlabDefCompiler:
         return f"{self.struct_name}.MESSAGE_HEADER = {self.struct_name}.typedefs.RTMA_MSG_HEADER;\n"
 
     def generate_hash_id(self, mdf: MDF) -> str:
-        return f'{self.struct_name}.hash.{self.sanitize_name(mdf.name)} = "{mdf.hash[:8]}";\n'
+        return f'{self.struct_name}.hash.{mdf.name} = "{mdf.hash[:8]}";\n'
 
     def generate_mex_opcodes(self) -> str:
         # these are copied from MatlabRTMA.h
