@@ -1,4 +1,4 @@
-"""This message def file was auto-generated by pyrtma.compile version 2.3.7"""
+"""This message def file was auto-generated by pyrtma.compile version 3.0.0"""
 
 import ctypes
 
@@ -34,14 +34,16 @@ if _ver_parse(__version__) >= _ver_parse("2.3.0"):
     from pyrtma.context import _update_context, get_context
 
 
-COMPILED_PYRTMA_VERSION: str = "2.3.7"
+COMPILED_PYRTMA_VERSION: str = "3.0.0"
 check_compiled_version(COMPILED_PYRTMA_VERSION)
 
 # Constants
-MAX_DATA_SETS: int = 6
+MAX_DATASETS: int = 6
+DATASET_NAME_LEN: int = 32
 MAX_LOGGER_FILENAME_LENGTH: int = 256
-MAX_MODULES: int = 200
-DYN_MOD_ID_START: int = 100
+MAX_MODULES: int = 256
+DYN_MOD_ID_START: int = 10000
+MAX_MODULE_ID: int = 32000
 MAX_HOSTS: int = 5
 MAX_MESSAGE_TYPES: int = 10000
 MAX_RTMA_MSG_TYPE: int = 99
@@ -74,27 +76,24 @@ MID_DATA_LOGGER: int = 4
 MID_QUICK_LOGGER: int = 5
 
 # Message Type IDs
-MT_DATA_LOGGER_STATUS: int = 62
-MT_DATA_LOGGER_STATUS_REQUEST: int = 63
-MT_ADD_DATA_COLLECTION: int = 64
-MT_REMOVE_DATA_COLLECTION: int = 65
-MT_ADD_DATA_SET: int = 66
-MT_REMOVE_DATA_SET: int = 67
-MT_DATA_COLLECTION_CONFIG_REQUEST: int = 68
-MT_DATA_COLLECTION_CONFIG: int = 69
-MT_DATA_COLLECTION_STARTED: int = 70
-MT_DATA_COLLECTION_STOPPED: int = 71
-MT_DATA_COLLECTION_SAVED: int = 72
-MT_DATA_LOGGER_START: int = 73
-MT_DATA_LOGGER_STOP: int = 74
-MT_DATA_LOGGER_PAUSE: int = 75
-MT_DATA_LOGGER_RESUME: int = 76
+MT_DATASET_STATUS_REQUEST: int = 62
+MT_DATASET_STATUS: int = 63
+MT_DATASET_ADD: int = 64
+MT_DATASET_ADDED: int = 65
+MT_DATASET_REMOVE: int = 66
+MT_DATASET_REMOVED: int = 67
+MT_DATA_LOGGER_CONFIG_REQUEST: int = 68
+MT_DATA_LOGGER_CONFIG: int = 69
+MT_DATASET_STARTED: int = 70
+MT_DATASET_STOPPED: int = 71
+MT_DATASET_SAVED: int = 72
+MT_DATASET_START: int = 73
+MT_DATASET_STOP: int = 74
+MT_DATASET_PAUSE: int = 75
+MT_DATASET_RESUME: int = 76
 MT_DATA_LOGGER_RESET: int = 77
 MT_DATA_LOGGER_ERROR: int = 78
-MT_DATA_LOGGER_METADATA_UPDATE: int = 79
-MT_DATA_LOGGER_METADATA_REQUEST: int = 87
-MT_DATA_LOGGER_METADATA: int = 88
-MT_DATA_LOG_TEST_2048: int = 89
+MT_DATA_LOG_TEST_2048: int = 79
 MT_LM_STATUS: int = 54
 MT_LM_EXIT: int = 55
 MT_SAVE_MESSAGE_LOG: int = 56
@@ -103,6 +102,7 @@ MT_PAUSE_MESSAGE_LOGGING: int = 58
 MT_RESUME_MESSAGE_LOGGING: int = 59
 MT_RESET_MESSAGE_LOG: int = 60
 MT_DUMP_MESSAGE_LOG: int = 61
+MT_DEBUG_TEXT: int = 91
 MT_LM_READY: int = 96
 MT_EXIT: int = 0
 MT_KILL: int = 1
@@ -114,12 +114,12 @@ MT_CONNECT: int = 13
 MT_DISCONNECT: int = 14
 MT_SUBSCRIBE: int = 15
 MT_UNSUBSCRIBE: int = 16
-MT_SHUTDOWN_RTMA: int = 17
 MT_MODULE_READY: int = 26
-MT_MESSAGE_TRAFFIC: int = 30
-MT_ACTIVE_CLIENTS: int = 31
-MT_CLIENT_INFO: int = 32
-MT_CLIENT_CLOSED: int = 33
+MT_PING: int = 29
+MT_PONG: int = 30
+MT_INTRODUCE: int = 31
+MT_HELLO: int = 32
+MT_GOODBYE: int = 33
 MT_CLIENT_SET_NAME: int = 34
 MT_RTMA_LOG: int = 40
 MT_RTMA_LOG_CRITICAL: int = 41
@@ -127,77 +127,27 @@ MT_RTMA_LOG_ERROR: int = 42
 MT_RTMA_LOG_WARNING: int = 43
 MT_RTMA_LOG_INFO: int = 44
 MT_RTMA_LOG_DEBUG: int = 45
-MT_TIMING_MESSAGE: int = 80
 MT_FORCE_DISCONNECT: int = 82
 MT_PAUSE_SUBSCRIPTION: int = 85
 MT_RESUME_SUBSCRIPTION: int = 86
 
 
 # Struct Definitions
-class DATA_SET(MessageBase, metaclass=MessageMeta):
-    type_name: ClassVar[str] = "DATA_SET"
-    type_hash: ClassVar[int] = 0xB267DE7B
-    type_size: ClassVar[int] = 452
+class DATASET(MessageBase, metaclass=MessageMeta):
+    type_name: ClassVar[str] = "DATASET"
+    type_hash: ClassVar[int] = 0xE2E58374
+    type_size: ClassVar[int] = 708
     type_source: ClassVar[str] = "data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_SET:\n  fields:\n    name: char[32]\n    sub_dir_fmt: char[128]\n    file_name_fmt: char[128]\n    formatter: char[32]\n    subdivide_interval: int32\n    msg_types: int32[32]'"
-    )
-
-    name: String = String(32)
-    sub_dir_fmt: String = String(128)
-    file_name_fmt: String = String(128)
-    formatter: String = String(32)
-    subdivide_interval: Int32 = Int32()
-    msg_types: IntArray[Int32] = IntArray(Int32, 32)
-
-
-class DATA_COLLECTION(MessageBase, metaclass=MessageMeta):
-    type_name: ClassVar[str] = "DATA_COLLECTION"
-    type_hash: ClassVar[int] = 0x9FA9C51A
-    type_size: ClassVar[int] = 3132
-    type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_COLLECTION:\n  fields:\n    name: char[32]\n    base_path: char[256]\n    dir_fmt: char[128]\n    num_data_sets: int16\n    unused: int16\n    data_sets: DATA_SET[MAX_DATA_SETS]'"
-    )
-
-    name: String = String(32)
-    base_path: String = String(256)
-    dir_fmt: String = String(128)
-    num_data_sets: Int16 = Int16()
-    unused: Int16 = Int16()
-    data_sets: StructArray[DATA_SET] = StructArray(DATA_SET, 6)
-
-
-class DATA_SET_INFO(MessageBase, metaclass=MessageMeta):
-    type_name: ClassVar[str] = "DATA_SET_INFO"
-    type_hash: ClassVar[int] = 0x65972A16
-    type_size: ClassVar[int] = 452
-    type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_SET_INFO:\n  fields:\n    name: char[32]\n    formatter: char[32]\n    save_path: char[256]\n    subdivide_interval: int32\n    msg_types: int32[32]'"
-    )
-
-    name: String = String(32)
-    formatter: String = String(32)
-    save_path: String = String(256)
-    subdivide_interval: Int32 = Int32()
-    msg_types: IntArray[Int32] = IntArray(Int32, 32)
-
-
-class DATA_COLLECTION_INFO(MessageBase, metaclass=MessageMeta):
-    type_name: ClassVar[str] = "DATA_COLLECTION_INFO"
-    type_hash: ClassVar[int] = 0x8FAF0CA3
-    type_size: ClassVar[int] = 3004
-    type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_COLLECTION_INFO:\n  fields:\n    name: char[32]\n    save_path: char[256]\n    num_data_sets: int16\n    unused: int16\n    data_sets: DATA_SET_INFO[MAX_DATA_SETS]'"
+        "'DATASET:\n  fields:\n    name: char[DATASET_NAME_LEN]\n    save_path: char[256]\n    filename: char[128]\n    formatter: char[32]\n    subdivide_interval: int32\n    msg_types: int32[64]'"
     )
 
     name: String = String(32)
     save_path: String = String(256)
-    num_data_sets: Int16 = Int16()
-    unused: Int16 = Int16()
-    data_sets: StructArray[DATA_SET_INFO] = StructArray(DATA_SET_INFO, 6)
+    filename: String = String(128)
+    formatter: String = String(32)
+    subdivide_interval: Int32 = Int32()
+    msg_types: IntArray[Int32] = IntArray(Int32, 64)
 
 
 class RTMA_MSG_HEADER(MessageBase, metaclass=MessageMeta):
@@ -225,16 +175,31 @@ class RTMA_MSG_HEADER(MessageBase, metaclass=MessageMeta):
 
 # Message Definitions
 @pyrtma.message_def
-class MDF_DATA_LOGGER_STATUS(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_STATUS_REQUEST(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 62
-    type_name: ClassVar[str] = "DATA_LOGGER_STATUS"
-    type_hash: ClassVar[int] = 0x9E922E93
-    type_size: ClassVar[int] = 24
+    type_name: ClassVar[str] = "DATASET_STATUS_REQUEST"
+    type_hash: ClassVar[int] = 0x2256A79F
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_LOGGER_STATUS:\n  id: 62\n  fields:\n    timestamp: double\n    elapsed_time: double\n    is_recording: int32\n    is_paused: int32'"
+        "'DATASET_STATUS_REQUEST:\n  id: 62\n  fields:\n    name: char[DATASET_NAME_LEN]'"
     )
 
+    name: String = String(32)
+
+
+@pyrtma.message_def
+class MDF_DATASET_STATUS(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 63
+    type_name: ClassVar[str] = "DATASET_STATUS"
+    type_hash: ClassVar[int] = 0xB608D5F6
+    type_size: ClassVar[int] = 56
+    type_source: ClassVar[str] = "data_logger.yaml"
+    type_def: ClassVar[str] = (
+        "'DATASET_STATUS:\n  id: 63\n  fields:\n    name: char[DATASET_NAME_LEN]\n    timestamp: double\n    elapsed_time: double\n    is_recording: int32\n    is_paused: int32'"
+    )
+
+    name: String = String(32)
     timestamp: Double = Double()
     elapsed_time: Double = Double()
     is_recording: Int32 = Int32()
@@ -242,175 +207,183 @@ class MDF_DATA_LOGGER_STATUS(MessageData, metaclass=MessageMeta):
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_STATUS_REQUEST(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 63
-    type_name: ClassVar[str] = "DATA_LOGGER_STATUS_REQUEST"
-    type_hash: ClassVar[int] = 0x168B0652
-    type_size: ClassVar[int] = 0
-    type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_LOGGER_STATUS_REQUEST:\n  id: 63\n  fields: null'"
-
-
-@pyrtma.message_def
-class MDF_ADD_DATA_COLLECTION(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_ADD(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 64
-    type_name: ClassVar[str] = "ADD_DATA_COLLECTION"
-    type_hash: ClassVar[int] = 0xA50C4BC8
-    type_size: ClassVar[int] = 3132
+    type_name: ClassVar[str] = "DATASET_ADD"
+    type_hash: ClassVar[int] = 0xD94DA73E
+    type_size: ClassVar[int] = 708
     type_source: ClassVar[str] = "data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'ADD_DATA_COLLECTION:\n  id: 64\n  fields:\n    collection: DATA_COLLECTION'"
+        "'DATASET_ADD:\n  id: 64\n  fields:\n    dataset: DATASET'"
     )
 
-    collection: Struct[DATA_COLLECTION] = Struct(DATA_COLLECTION)
+    dataset: Struct[DATASET] = Struct(DATASET)
 
 
 @pyrtma.message_def
-class MDF_REMOVE_DATA_COLLECTION(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_ADDED(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 65
-    type_name: ClassVar[str] = "REMOVE_DATA_COLLECTION"
-    type_hash: ClassVar[int] = 0x5430DCF2
+    type_name: ClassVar[str] = "DATASET_ADDED"
+    type_hash: ClassVar[int] = 0xCFA9917E
     type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'REMOVE_DATA_COLLECTION:\n  id: 65\n  fields:\n    collection_name: char[32]'"
+        "'DATASET_ADDED:\n  id: 65\n  fields:\n    name: char[DATASET_NAME_LEN]'"
     )
 
-    collection_name: String = String(32)
-
-
-@pyrtma.message_def
-class MDF_ADD_DATA_SET(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 66
-    type_name: ClassVar[str] = "ADD_DATA_SET"
-    type_hash: ClassVar[int] = 0xB368D74F
-    type_size: ClassVar[int] = 484
-    type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'ADD_DATA_SET:\n  id: 66\n  fields:\n    collection_name: char[32]\n    data_set: DATA_SET'"
-    )
-
-    collection_name: String = String(32)
-    data_set: Struct[DATA_SET] = Struct(DATA_SET)
-
-
-@pyrtma.message_def
-class MDF_REMOVE_DATA_SET(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 67
-    type_name: ClassVar[str] = "REMOVE_DATA_SET"
-    type_hash: ClassVar[int] = 0x0BC39C04
-    type_size: ClassVar[int] = 64
-    type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'REMOVE_DATA_SET:\n  id: 67\n  fields:\n    collection_name: char[32]\n    name: char[32]'"
-    )
-
-    collection_name: String = String(32)
     name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_COLLECTION_CONFIG_REQUEST(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_REMOVE(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 66
+    type_name: ClassVar[str] = "DATASET_REMOVE"
+    type_hash: ClassVar[int] = 0x13EEE304
+    type_size: ClassVar[int] = 32
+    type_source: ClassVar[str] = "data_logger.yaml"
+    type_def: ClassVar[str] = (
+        "'DATASET_REMOVE:\n  id: 66\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
+
+
+@pyrtma.message_def
+class MDF_DATASET_REMOVED(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 67
+    type_name: ClassVar[str] = "DATASET_REMOVED"
+    type_hash: ClassVar[int] = 0x7F388965
+    type_size: ClassVar[int] = 32
+    type_source: ClassVar[str] = "data_logger.yaml"
+    type_def: ClassVar[str] = (
+        "'DATASET_REMOVED:\n  id: 67\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
+
+
+@pyrtma.message_def
+class MDF_DATA_LOGGER_CONFIG_REQUEST(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 68
-    type_name: ClassVar[str] = "DATA_COLLECTION_CONFIG_REQUEST"
-    type_hash: ClassVar[int] = 0x1779D826
+    type_name: ClassVar[str] = "DATA_LOGGER_CONFIG_REQUEST"
+    type_hash: ClassVar[int] = 0x91598F0A
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_COLLECTION_CONFIG_REQUEST:\n  id: 68\n  fields: null'"
-    )
+    type_def: ClassVar[str] = "'DATA_LOGGER_CONFIG_REQUEST:\n  id: 68\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF_DATA_COLLECTION_CONFIG(MessageData, metaclass=MessageMeta):
+class MDF_DATA_LOGGER_CONFIG(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 69
-    type_name: ClassVar[str] = "DATA_COLLECTION_CONFIG"
-    type_hash: ClassVar[int] = 0xC08D3576
-    type_size: ClassVar[int] = 3132
+    type_name: ClassVar[str] = "DATA_LOGGER_CONFIG"
+    type_hash: ClassVar[int] = 0x085048FB
+    type_size: ClassVar[int] = 4252
     type_source: ClassVar[str] = "data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_COLLECTION_CONFIG:\n  id: 69\n  fields:\n    collection: DATA_COLLECTION'"
+        "'DATA_LOGGER_CONFIG:\n  id: 69\n  fields:\n    num_datasets: int32\n    datasets: DATASET[6]'"
     )
 
-    collection: Struct[DATA_COLLECTION] = Struct(DATA_COLLECTION)
+    num_datasets: Int32 = Int32()
+    datasets: StructArray[DATASET] = StructArray(DATASET, 6)
 
 
 @pyrtma.message_def
-class MDF_DATA_COLLECTION_STARTED(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_STARTED(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 70
-    type_name: ClassVar[str] = "DATA_COLLECTION_STARTED"
-    type_hash: ClassVar[int] = 0x415737CA
-    type_size: ClassVar[int] = 3004
+    type_name: ClassVar[str] = "DATASET_STARTED"
+    type_hash: ClassVar[int] = 0x5A6BC4E6
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_COLLECTION_STARTED:\n  id: 70\n  fields:\n    collection: DATA_COLLECTION_INFO'"
+        "'DATASET_STARTED:\n  id: 70\n  fields:\n    name: char[DATASET_NAME_LEN]'"
     )
 
-    collection: Struct[DATA_COLLECTION_INFO] = Struct(DATA_COLLECTION_INFO)
+    name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_COLLECTION_STOPPED(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_STOPPED(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 71
-    type_name: ClassVar[str] = "DATA_COLLECTION_STOPPED"
-    type_hash: ClassVar[int] = 0xDA1AF3AE
-    type_size: ClassVar[int] = 3004
+    type_name: ClassVar[str] = "DATASET_STOPPED"
+    type_hash: ClassVar[int] = 0x19F8847F
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_COLLECTION_STOPPED:\n  id: 71\n  fields:\n    collection: DATA_COLLECTION_INFO'"
+        "'DATASET_STOPPED:\n  id: 71\n  fields:\n    name: char[DATASET_NAME_LEN]'"
     )
 
-    collection: Struct[DATA_COLLECTION_INFO] = Struct(DATA_COLLECTION_INFO)
+    name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_COLLECTION_SAVED(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_SAVED(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 72
-    type_name: ClassVar[str] = "DATA_COLLECTION_SAVED"
-    type_hash: ClassVar[int] = 0x74528518
-    type_size: ClassVar[int] = 0
+    type_name: ClassVar[str] = "DATASET_SAVED"
+    type_hash: ClassVar[int] = 0x964F725E
+    type_size: ClassVar[int] = 1056
     type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_COLLECTION_SAVED:\n  id: 72\n  fields: null'"
+    type_def: ClassVar[str] = (
+        "'DATASET_SAVED:\n  id: 72\n  fields:\n    name: char[DATASET_NAME_LEN]\n    filepath: char[1024]'"
+    )
+
+    name: String = String(32)
+    filepath: String = String(1024)
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_START(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_START(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 73
-    type_name: ClassVar[str] = "DATA_LOGGER_START"
-    type_hash: ClassVar[int] = 0x982D6942
-    type_size: ClassVar[int] = 0
+    type_name: ClassVar[str] = "DATASET_START"
+    type_hash: ClassVar[int] = 0x019BBFB4
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_LOGGER_START:\n  id: 73\n  fields: null'"
+    type_def: ClassVar[str] = (
+        "'DATASET_START:\n  id: 73\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_STOP(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_STOP(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 74
-    type_name: ClassVar[str] = "DATA_LOGGER_STOP"
-    type_hash: ClassVar[int] = 0xA5B66B24
-    type_size: ClassVar[int] = 0
+    type_name: ClassVar[str] = "DATASET_STOP"
+    type_hash: ClassVar[int] = 0xFC4A83B1
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_LOGGER_STOP:\n  id: 74\n  fields: null'"
+    type_def: ClassVar[str] = (
+        "'DATASET_STOP:\n  id: 74\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_PAUSE(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_PAUSE(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 75
-    type_name: ClassVar[str] = "DATA_LOGGER_PAUSE"
-    type_hash: ClassVar[int] = 0x9A556835
-    type_size: ClassVar[int] = 0
+    type_name: ClassVar[str] = "DATASET_PAUSE"
+    type_hash: ClassVar[int] = 0x0DCFA7B0
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_LOGGER_PAUSE:\n  id: 75\n  fields: null'"
+    type_def: ClassVar[str] = (
+        "'DATASET_PAUSE:\n  id: 75\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_RESUME(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_RESUME(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 76
-    type_name: ClassVar[str] = "DATA_LOGGER_RESUME"
-    type_hash: ClassVar[int] = 0x18811751
-    type_size: ClassVar[int] = 0
+    type_name: ClassVar[str] = "DATASET_RESUME"
+    type_hash: ClassVar[int] = 0xF4E01F27
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_LOGGER_RESUME:\n  id: 76\n  fields: null'"
+    type_def: ClassVar[str] = (
+        "'DATASET_RESUME:\n  id: 76\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
 
 
 @pyrtma.message_def
@@ -427,65 +400,27 @@ class MDF_DATA_LOGGER_RESET(MessageData, metaclass=MessageMeta):
 class MDF_DATA_LOGGER_ERROR(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 78
     type_name: ClassVar[str] = "DATA_LOGGER_ERROR"
-    type_hash: ClassVar[int] = 0xAD3B4A67
-    type_size: ClassVar[int] = 512
+    type_hash: ClassVar[int] = 0x630C7107
+    type_size: ClassVar[int] = 608
     type_source: ClassVar[str] = "data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_LOGGER_ERROR:\n  id: 78\n  fields:\n    msg: char[512]'"
+        "'DATA_LOGGER_ERROR:\n  id: 78\n  fields:\n    dataset_name: char[DATASET_NAME_LEN]\n    exc_type: char[64]\n    msg: char[512]'"
     )
 
+    dataset_name: String = String(32)
+    exc_type: String = String(64)
     msg: String = String(512)
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_METADATA_UPDATE(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 79
-    type_name: ClassVar[str] = "DATA_LOGGER_METADATA_UPDATE"
-    type_hash: ClassVar[int] = 0x3C673526
-    type_size: ClassVar[int] = 512
-    type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_LOGGER_METADATA_UPDATE:\n  id: 79\n  fields:\n    json: char[512]'"
-    )
-
-    json: String = String(512)
-
-
-@pyrtma.message_def
-class MDF_DATA_LOGGER_METADATA_REQUEST(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 87
-    type_name: ClassVar[str] = "DATA_LOGGER_METADATA_REQUEST"
-    type_hash: ClassVar[int] = 0x2D068003
-    type_size: ClassVar[int] = 0
-    type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_LOGGER_METADATA_REQUEST:\n  id: 87\n  fields: null'"
-    )
-
-
-@pyrtma.message_def
-class MDF_DATA_LOGGER_METADATA(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 88
-    type_name: ClassVar[str] = "DATA_LOGGER_METADATA"
-    type_hash: ClassVar[int] = 0x3852789B
-    type_size: ClassVar[int] = 1024
-    type_source: ClassVar[str] = "data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_LOGGER_METADATA:\n  id: 88\n  fields:\n    json: char[1024]'"
-    )
-
-    json: String = String(1024)
-
-
-@pyrtma.message_def
 class MDF_DATA_LOG_TEST_2048(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 89
+    type_id: ClassVar[int] = 79
     type_name: ClassVar[str] = "DATA_LOG_TEST_2048"
-    type_hash: ClassVar[int] = 0x651D3C98
+    type_hash: ClassVar[int] = 0x6F24B1D1
     type_size: ClassVar[int] = 2048
     type_source: ClassVar[str] = "data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_LOG_TEST_2048:\n  id: 89\n  fields:\n    raw: char[2048]'"
+        "'DATA_LOG_TEST_2048:\n  id: 79\n  fields:\n    raw: char[2048]'"
     )
 
     raw: String = String(2048)
@@ -585,6 +520,16 @@ class MDF_DUMP_MESSAGE_LOG(MessageData, metaclass=MessageMeta):
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "quick_logger.yaml"
     type_def: ClassVar[str] = "'DUMP_MESSAGE_LOG:\n  id: 61\n  fields: null'"
+
+
+@pyrtma.message_def
+class MDF_DEBUG_TEXT(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 91
+    type_name: ClassVar[str] = "DEBUG_TEXT"
+    type_hash: ClassVar[int] = 0xEC89E79E
+    type_size: ClassVar[int] = 0
+    type_source: ClassVar[str] = "quick_logger.yaml"
+    type_def: ClassVar[str] = "'DEBUG_TEXT:\n  id: 91\n  fields: null'"
 
 
 @pyrtma.message_def
@@ -733,16 +678,6 @@ class MDF_UNSUBSCRIBE(MessageData, metaclass=MessageMeta):
 
 
 @pyrtma.message_def
-class MDF_SHUTDOWN_RTMA(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 17
-    type_name: ClassVar[str] = "SHUTDOWN_RTMA"
-    type_hash: ClassVar[int] = 0x0912BFA5
-    type_size: ClassVar[int] = 0
-    type_source: ClassVar[str] = "core_defs.yaml"
-    type_def: ClassVar[str] = "'SHUTDOWN_RTMA:\n  id: 17\n  fields: null'"
-
-
-@pyrtma.message_def
 class MDF_MODULE_READY(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 26
     type_name: ClassVar[str] = "MODULE_READY"
@@ -755,82 +690,74 @@ class MDF_MODULE_READY(MessageData, metaclass=MessageMeta):
 
 
 @pyrtma.message_def
-class MDF_MESSAGE_TRAFFIC(MessageData, metaclass=MessageMeta):
+class MDF_PING(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 29
+    type_name: ClassVar[str] = "PING"
+    type_hash: ClassVar[int] = 0x43014E6F
+    type_size: ClassVar[int] = 4
+    type_source: ClassVar[str] = "core_defs.yaml"
+    type_def: ClassVar[str] = "'PING:\n  id: 29\n  fields:\n    uid: int32'"
+
+    uid: Int32 = Int32()
+
+
+@pyrtma.message_def
+class MDF_PONG(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 30
-    type_name: ClassVar[str] = "MESSAGE_TRAFFIC"
-    type_hash: ClassVar[int] = 0xA5D05FF5
-    type_size: ClassVar[int] = 408
+    type_name: ClassVar[str] = "PONG"
+    type_hash: ClassVar[int] = 0xDAC22B06
+    type_size: ClassVar[int] = 4
     type_source: ClassVar[str] = "core_defs.yaml"
-    type_def: ClassVar[str] = (
-        "'MESSAGE_TRAFFIC:\n  id: 30\n  fields:\n    seqno: uint32\n    sub_seqno: uint32\n    start_timestamp: double\n    end_timestamp: double\n    msg_type: MSG_TYPE[MESSAGE_TRAFFIC_SIZE]\n    msg_count: uint16[MESSAGE_TRAFFIC_SIZE]'"
-    )
+    type_def: ClassVar[str] = "'PONG:\n  id: 30\n  fields:\n    uid: int32'"
 
-    seqno: Uint32 = Uint32()
-    sub_seqno: Uint32 = Uint32()
-    start_timestamp: Double = Double()
-    end_timestamp: Double = Double()
-    msg_type: IntArray[Int32] = IntArray(Int32, 64)
-    msg_count: IntArray[Uint16] = IntArray(Uint16, 64)
+    uid: Int32 = Int32()
 
 
 @pyrtma.message_def
-class MDF_ACTIVE_CLIENTS(MessageData, metaclass=MessageMeta):
+class MDF_INTRODUCE(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 31
-    type_name: ClassVar[str] = "ACTIVE_CLIENTS"
-    type_hash: ClassVar[int] = 0xFC42B3AF
-    type_size: ClassVar[int] = 1552
+    type_name: ClassVar[str] = "INTRODUCE"
+    type_hash: ClassVar[int] = 0x6F4769E8
+    type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "core_defs.yaml"
-    type_def: ClassVar[str] = (
-        "'ACTIVE_CLIENTS:\n  id: 31\n  fields:\n    timestamp: double\n    num_clients: int16\n    padding: int16\n    reserved: int32\n    client_mod_id: MODULE_ID[MAX_ACTIVE_CLIENTS]\n    client_pid: int32[MAX_ACTIVE_CLIENTS]'"
-    )
-
-    timestamp: Double = Double()
-    num_clients: Int16 = Int16()
-    padding: Int16 = Int16()
-    reserved: Int32 = Int32()
-    client_mod_id: IntArray[Int16] = IntArray(Int16, 256)
-    client_pid: IntArray[Int32] = IntArray(Int32, 256)
+    type_def: ClassVar[str] = "'INTRODUCE:\n  id: 31\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF_CLIENT_INFO(MessageData, metaclass=MessageMeta):
+class MDF_HELLO(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 32
-    type_name: ClassVar[str] = "CLIENT_INFO"
-    type_hash: ClassVar[int] = 0x401076E7
-    type_size: ClassVar[int] = 80
+    type_name: ClassVar[str] = "HELLO"
+    type_hash: ClassVar[int] = 0xF393D643
+    type_size: ClassVar[int] = 76
     type_source: ClassVar[str] = "core_defs.yaml"
     type_def: ClassVar[str] = (
-        "'CLIENT_INFO:\n  id: 32\n  fields:\n    addr: char[32]\n    uid: int32\n    pid: int32\n    mod_id: MODULE_ID\n    is_logger: int16\n    is_unique: int16\n    port: uint16\n    name: char[MAX_NAME_LEN]'"
+        "'HELLO:\n  id: 32\n  fields:\n    uid: int32\n    pid: int32\n    mod_id: MODULE_ID\n    port: uint16\n    addr: char[MAX_NAME_LEN]\n    name: char[MAX_NAME_LEN]'"
     )
 
-    addr: String = String(32)
     uid: Int32 = Int32()
     pid: Int32 = Int32()
     mod_id: Int16 = Int16()
-    is_logger: Int16 = Int16()
-    is_unique: Int16 = Int16()
     port: Uint16 = Uint16()
+    addr: String = String(32)
     name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_CLIENT_CLOSED(MessageData, metaclass=MessageMeta):
+class MDF_GOODBYE(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 33
-    type_name: ClassVar[str] = "CLIENT_CLOSED"
-    type_hash: ClassVar[int] = 0x087F4249
-    type_size: ClassVar[int] = 80
+    type_name: ClassVar[str] = "GOODBYE"
+    type_hash: ClassVar[int] = 0x1EC8D03F
+    type_size: ClassVar[int] = 76
     type_source: ClassVar[str] = "core_defs.yaml"
     type_def: ClassVar[str] = (
-        "'CLIENT_CLOSED:\n  id: 33\n  fields:\n    addr: char[32]\n    uid: int32\n    pid: int32\n    mod_id: MODULE_ID\n    is_logger: int16\n    is_unique: int16\n    port: uint16\n    name: char[MAX_NAME_LEN]'"
+        "'GOODBYE:\n  id: 33\n  fields:\n    uid: int32\n    pid: int32\n    mod_id: MODULE_ID\n    port: uint16\n    addr: char[MAX_NAME_LEN]\n    name: char[MAX_NAME_LEN]'"
     )
 
-    addr: String = String(32)
     uid: Int32 = Int32()
     pid: Int32 = Int32()
     mod_id: Int16 = Int16()
-    is_logger: Int16 = Int16()
-    is_unique: Int16 = Int16()
     port: Uint16 = Uint16()
+    addr: String = String(32)
     name: String = String(32)
 
 
@@ -966,22 +893,6 @@ class MDF_RTMA_LOG_DEBUG(MessageData, metaclass=MessageMeta):
     pathname: String = String(512)
     funcname: String = String(256)
     message: String = String(1024)
-
-
-@pyrtma.message_def
-class MDF_TIMING_MESSAGE(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 80
-    type_name: ClassVar[str] = "TIMING_MESSAGE"
-    type_hash: ClassVar[int] = 0xDAA7503D
-    type_size: ClassVar[int] = 20808
-    type_source: ClassVar[str] = "core_defs.yaml"
-    type_def: ClassVar[str] = (
-        "'TIMING_MESSAGE:\n  id: 80\n  fields:\n    timing: uint16[MAX_MESSAGE_TYPES]\n    ModulePID: int32[MAX_MODULES]\n    send_time: double'"
-    )
-
-    timing: IntArray[Uint16] = IntArray(Uint16, 10000)
-    ModulePID: IntArray[Int32] = IntArray(Int32, 200)
-    send_time: Double = Double()
 
 
 @pyrtma.message_def
