@@ -1,3 +1,5 @@
+"""data_logger Json Formatter"""
+
 import pyrtma
 
 from ..data_formatter import DataFormatter
@@ -7,7 +9,7 @@ from typing import ClassVar
 class JsonFormatter(DataFormatter):
     name: ClassVar[str] = "json"
     mode: ClassVar[str] = "wt"
-    ext: ClassVar[str] = ".json"
+    ext: ClassVar[str] = ".jsonl"
 
     def format_message(self, msg: pyrtma.Message) -> str:
         return msg.to_json(minify=True) + "\n"
