@@ -1,3 +1,5 @@
+"""data_logger Raw Formatter"""
+
 import pyrtma
 
 from ..data_formatter import DataFormatter
@@ -7,7 +9,7 @@ from typing import ClassVar
 class RawFormatter(DataFormatter):
     name: ClassVar[str] = "raw"
     mode: ClassVar[str] = "wb"
-    ext: ClassVar[str] = ".raw"
+    ext: ClassVar[str] = ".dat"
 
     def format_message(self, msg: pyrtma.Message) -> bytes:
         return bytes(msg.header) + bytes(msg.data)
