@@ -1,4 +1,4 @@
-"""This message def file was auto-generated by pyrtma.compile version 2.3.7"""
+"""This message def file was auto-generated by pyrtma.compile version 3.0.0"""
 
 import ctypes
 
@@ -34,11 +34,12 @@ if _ver_parse(__version__) >= _ver_parse("2.3.0"):
     from pyrtma.context import _update_context, get_context
 
 
-COMPILED_PYRTMA_VERSION: str = "2.3.7"
+COMPILED_PYRTMA_VERSION: str = "3.0.0"
 check_compiled_version(COMPILED_PYRTMA_VERSION)
 
 # Constants
-MAX_DATA_SETS: int = 6
+MAX_DATASETS: int = 6
+DATASET_NAME_LEN: int = 32
 MAX_LOGGER_FILENAME_LENGTH: int = 256
 MAX_MODULES: int = 200
 DYN_MOD_ID_START: int = 100
@@ -208,27 +209,24 @@ MID_SENSE_TOUCH_INTERFACE: int = 98
 MID_SENSOR_STIM_TRANSFORM_PY: int = 99
 
 # Message Type IDs
-MT_DATA_LOGGER_STATUS: int = 62
-MT_DATA_LOGGER_STATUS_REQUEST: int = 63
-MT_ADD_DATA_COLLECTION: int = 64
-MT_REMOVE_DATA_COLLECTION: int = 65
-MT_ADD_DATA_SET: int = 66
-MT_REMOVE_DATA_SET: int = 67
-MT_DATA_COLLECTION_CONFIG_REQUEST: int = 68
-MT_DATA_COLLECTION_CONFIG: int = 69
-MT_DATA_COLLECTION_STARTED: int = 70
-MT_DATA_COLLECTION_STOPPED: int = 71
-MT_DATA_COLLECTION_SAVED: int = 72
-MT_DATA_LOGGER_START: int = 73
-MT_DATA_LOGGER_STOP: int = 74
-MT_DATA_LOGGER_PAUSE: int = 75
-MT_DATA_LOGGER_RESUME: int = 76
+MT_DATASET_STATUS_REQUEST: int = 62
+MT_DATASET_STATUS: int = 63
+MT_DATASET_ADD: int = 64
+MT_DATASET_ADDED: int = 65
+MT_DATASET_REMOVE: int = 66
+MT_DATASET_REMOVED: int = 67
+MT_DATA_LOGGER_CONFIG_REQUEST: int = 68
+MT_DATA_LOGGER_CONFIG: int = 69
+MT_DATASET_STARTED: int = 70
+MT_DATASET_STOPPED: int = 71
+MT_DATASET_SAVED: int = 72
+MT_DATASET_START: int = 73
+MT_DATASET_STOP: int = 74
+MT_DATASET_PAUSE: int = 75
+MT_DATASET_RESUME: int = 76
 MT_DATA_LOGGER_RESET: int = 77
 MT_DATA_LOGGER_ERROR: int = 78
-MT_DATA_LOGGER_METADATA_UPDATE: int = 79
-MT_DATA_LOGGER_METADATA_REQUEST: int = 87
-MT_DATA_LOGGER_METADATA: int = 88
-MT_DATA_LOG_TEST_2048: int = 89
+MT_DATA_LOG_TEST_2048: int = 79
 MT_LM_STATUS: int = 54
 MT_LM_EXIT: int = 55
 MT_SAVE_MESSAGE_LOG: int = 56
@@ -237,23 +235,26 @@ MT_PAUSE_MESSAGE_LOGGING: int = 58
 MT_RESUME_MESSAGE_LOGGING: int = 59
 MT_RESET_MESSAGE_LOG: int = 60
 MT_DUMP_MESSAGE_LOG: int = 61
+MT_DEBUG_TEXT: int = 91
 MT_LM_READY: int = 96
 MT_EXIT: int = 0
 MT_KILL: int = 1
 MT_ACKNOWLEDGE: int = 2
 MT_CONNECT_V2: int = 4
 MT_FAIL_SUBSCRIBE: int = 6
+MT_SUBSCRIPTION_OPTION: int = 7
 MT_FAILED_MESSAGE: int = 8
+MT_MESSAGE_TIMING: int = 9
 MT_CONNECT: int = 13
 MT_DISCONNECT: int = 14
 MT_SUBSCRIBE: int = 15
 MT_UNSUBSCRIBE: int = 16
-MT_SHUTDOWN_RTMA: int = 17
 MT_MODULE_READY: int = 26
-MT_MESSAGE_TRAFFIC: int = 30
-MT_ACTIVE_CLIENTS: int = 31
-MT_CLIENT_INFO: int = 32
-MT_CLIENT_CLOSED: int = 33
+MT_PING: int = 29
+MT_PONG: int = 30
+MT_INTRODUCE: int = 31
+MT_HELLO: int = 32
+MT_GOODBYE: int = 33
 MT_CLIENT_SET_NAME: int = 34
 MT_RTMA_LOG: int = 40
 MT_RTMA_LOG_CRITICAL: int = 41
@@ -467,26 +468,26 @@ MT_MECH_STIM_CANCEL: int = 4265
 MT_MECH_STIM_DONE: int = 4266
 MT_MECH_STIM_ERROR: int = 4267
 MT_UC_MECH_STIM_CONFIGURE: int = 4268
-MT__RESERVED_004300: int = 4300
-MT__RESERVED_004360: int = 4360
-MT__RESERVED_004361: int = 4361
-MT__RESERVED_004362: int = 4362
-MT__RESERVED_004363: int = 4363
-MT__RESERVED_004364: int = 4364
-MT__RESERVED_004365: int = 4365
-MT__RESERVED_004366: int = 4366
-MT__RESERVED_004367: int = 4367
-MT__RESERVED_004368: int = 4368
-MT__RESERVED_004369: int = 4369
-MT__RESERVED_004370: int = 4370
-MT__RESERVED_004371: int = 4371
-MT__RESERVED_004372: int = 4372
-MT__RESERVED_004373: int = 4373
-MT__RESERVED_004374: int = 4374
-MT__RESERVED_004375: int = 4375
-MT__RESERVED_004376: int = 4376
-MT__RESERVED_004377: int = 4377
-MT__RESERVED_004378: int = 4378
+MT_RESERVED_004300: int = 4300
+MT_RESERVED_004360: int = 4360
+MT_RESERVED_004361: int = 4361
+MT_RESERVED_004362: int = 4362
+MT_RESERVED_004363: int = 4363
+MT_RESERVED_004364: int = 4364
+MT_RESERVED_004365: int = 4365
+MT_RESERVED_004366: int = 4366
+MT_RESERVED_004367: int = 4367
+MT_RESERVED_004368: int = 4368
+MT_RESERVED_004369: int = 4369
+MT_RESERVED_004370: int = 4370
+MT_RESERVED_004371: int = 4371
+MT_RESERVED_004372: int = 4372
+MT_RESERVED_004373: int = 4373
+MT_RESERVED_004374: int = 4374
+MT_RESERVED_004375: int = 4375
+MT_RESERVED_004376: int = 4376
+MT_RESERVED_004377: int = 4377
+MT_RESERVED_004378: int = 4378
 MT_VALIDATOR_A: int = 5000
 MT_VALIDATOR_B: int = 5001
 MT_TEST_MSG_128: int = 5002
@@ -500,73 +501,25 @@ MT_SUBSCRIBER_READY: int = 5009
 MT_PUBLISHER_READY: int = 5010
 MT_SUBSCRIBER_DONE: int = 5011
 MT_PUBLISHER_DONE: int = 5012
+MT_ARRAY_OF_ONE: int = 5013
 
 
 # Struct Definitions
-class DATA_SET(MessageBase, metaclass=MessageMeta):
-    type_name: ClassVar[str] = "DATA_SET"
-    type_hash: ClassVar[int] = 0xB267DE7B
-    type_size: ClassVar[int] = 452
+class DATASET(MessageBase, metaclass=MessageMeta):
+    type_name: ClassVar[str] = "DATASET"
+    type_hash: ClassVar[int] = 0xE2E58374
+    type_size: ClassVar[int] = 708
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_SET:\n  fields:\n    name: char[32]\n    sub_dir_fmt: char[128]\n    file_name_fmt: char[128]\n    formatter: char[32]\n    subdivide_interval: int32\n    msg_types: int32[32]'"
-    )
-
-    name: String = String(32)
-    sub_dir_fmt: String = String(128)
-    file_name_fmt: String = String(128)
-    formatter: String = String(32)
-    subdivide_interval: Int32 = Int32()
-    msg_types: IntArray[Int32] = IntArray(Int32, 32)
-
-
-class DATA_COLLECTION(MessageBase, metaclass=MessageMeta):
-    type_name: ClassVar[str] = "DATA_COLLECTION"
-    type_hash: ClassVar[int] = 0x9FA9C51A
-    type_size: ClassVar[int] = 3132
-    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_COLLECTION:\n  fields:\n    name: char[32]\n    base_path: char[256]\n    dir_fmt: char[128]\n    num_data_sets: int16\n    unused: int16\n    data_sets: DATA_SET[MAX_DATA_SETS]'"
-    )
-
-    name: String = String(32)
-    base_path: String = String(256)
-    dir_fmt: String = String(128)
-    num_data_sets: Int16 = Int16()
-    unused: Int16 = Int16()
-    data_sets: StructArray[DATA_SET] = StructArray(DATA_SET, 6)
-
-
-class DATA_SET_INFO(MessageBase, metaclass=MessageMeta):
-    type_name: ClassVar[str] = "DATA_SET_INFO"
-    type_hash: ClassVar[int] = 0x65972A16
-    type_size: ClassVar[int] = 452
-    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_SET_INFO:\n  fields:\n    name: char[32]\n    formatter: char[32]\n    save_path: char[256]\n    subdivide_interval: int32\n    msg_types: int32[32]'"
-    )
-
-    name: String = String(32)
-    formatter: String = String(32)
-    save_path: String = String(256)
-    subdivide_interval: Int32 = Int32()
-    msg_types: IntArray[Int32] = IntArray(Int32, 32)
-
-
-class DATA_COLLECTION_INFO(MessageBase, metaclass=MessageMeta):
-    type_name: ClassVar[str] = "DATA_COLLECTION_INFO"
-    type_hash: ClassVar[int] = 0x8FAF0CA3
-    type_size: ClassVar[int] = 3004
-    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_COLLECTION_INFO:\n  fields:\n    name: char[32]\n    save_path: char[256]\n    num_data_sets: int16\n    unused: int16\n    data_sets: DATA_SET_INFO[MAX_DATA_SETS]'"
+        "'DATASET:\n  fields:\n    name: char[DATASET_NAME_LEN]\n    save_path: char[256]\n    filename: char[128]\n    formatter: char[32]\n    subdivide_interval: int32\n    msg_types: int32[64]'"
     )
 
     name: String = String(32)
     save_path: String = String(256)
-    num_data_sets: Int16 = Int16()
-    unused: Int16 = Int16()
-    data_sets: StructArray[DATA_SET_INFO] = StructArray(DATA_SET_INFO, 6)
+    filename: String = String(128)
+    formatter: String = String(32)
+    subdivide_interval: Int32 = Int32()
+    msg_types: IntArray[Int32] = IntArray(Int32, 64)
 
 
 class RTMA_MSG_HEADER(MessageBase, metaclass=MessageMeta):
@@ -620,24 +573,24 @@ class VALIDATOR_STRUCT(MessageBase, metaclass=MessageMeta):
     int32: Int32 = Int32()
     int64: Int64 = Int64()
     uint8: Uint8 = Uint8()
-    padding_0_: Char = Char()
+    padding_0_: ByteArray = ByteArray(1)
     uint16: Uint16 = Uint16()
     uint32: Uint32 = Uint32()
     uint64: Uint64 = Uint64()
     int8_arr: IntArray[Int8] = IntArray(Int8, 4)
     int16_arr: IntArray[Int16] = IntArray(Int16, 4)
     int32_arr: IntArray[Int32] = IntArray(Int32, 4)
-    padding_1_: String = String(4)
+    padding_1_: ByteArray = ByteArray(4)
     int64_arr: IntArray[Int64] = IntArray(Int64, 4)
     uint8_arr: IntArray[Uint8] = IntArray(Uint8, 4)
     uint16_arr: IntArray[Uint16] = IntArray(Uint16, 4)
     uint32_arr: IntArray[Uint32] = IntArray(Uint32, 4)
-    padding_2_: String = String(4)
+    padding_2_: ByteArray = ByteArray(4)
     uint64_arr: IntArray[Uint64] = IntArray(Uint64, 4)
     byte_: Byte = Byte()
     byte_arr: ByteArray = ByteArray(4)
     string: String = String(4)
-    padding_3_: String = String(3)
+    padding_3_: ByteArray = ByteArray(3)
     float_: Float = Float()
     double_: Double = Double()
     float_arr: FloatArray[Float] = FloatArray(Float, 4)
@@ -747,16 +700,31 @@ class DYNAMIXEL_INFO(MessageBase, metaclass=MessageMeta):
 
 # Message Definitions
 @pyrtma.message_def
-class MDF_DATA_LOGGER_STATUS(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_STATUS_REQUEST(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 62
-    type_name: ClassVar[str] = "DATA_LOGGER_STATUS"
-    type_hash: ClassVar[int] = 0x9E922E93
-    type_size: ClassVar[int] = 24
+    type_name: ClassVar[str] = "DATASET_STATUS_REQUEST"
+    type_hash: ClassVar[int] = 0x2256A79F
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_LOGGER_STATUS:\n  id: 62\n  fields:\n    timestamp: double\n    elapsed_time: double\n    is_recording: int32\n    is_paused: int32'"
+        "'DATASET_STATUS_REQUEST:\n  id: 62\n  fields:\n    name: char[DATASET_NAME_LEN]'"
     )
 
+    name: String = String(32)
+
+
+@pyrtma.message_def
+class MDF_DATASET_STATUS(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 63
+    type_name: ClassVar[str] = "DATASET_STATUS"
+    type_hash: ClassVar[int] = 0xB608D5F6
+    type_size: ClassVar[int] = 56
+    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
+    type_def: ClassVar[str] = (
+        "'DATASET_STATUS:\n  id: 63\n  fields:\n    name: char[DATASET_NAME_LEN]\n    timestamp: double\n    elapsed_time: double\n    is_recording: int32\n    is_paused: int32'"
+    )
+
+    name: String = String(32)
     timestamp: Double = Double()
     elapsed_time: Double = Double()
     is_recording: Int32 = Int32()
@@ -764,175 +732,183 @@ class MDF_DATA_LOGGER_STATUS(MessageData, metaclass=MessageMeta):
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_STATUS_REQUEST(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 63
-    type_name: ClassVar[str] = "DATA_LOGGER_STATUS_REQUEST"
-    type_hash: ClassVar[int] = 0x168B0652
-    type_size: ClassVar[int] = 0
-    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_LOGGER_STATUS_REQUEST:\n  id: 63\n  fields: null'"
-
-
-@pyrtma.message_def
-class MDF_ADD_DATA_COLLECTION(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_ADD(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 64
-    type_name: ClassVar[str] = "ADD_DATA_COLLECTION"
-    type_hash: ClassVar[int] = 0xA50C4BC8
-    type_size: ClassVar[int] = 3132
+    type_name: ClassVar[str] = "DATASET_ADD"
+    type_hash: ClassVar[int] = 0xD94DA73E
+    type_size: ClassVar[int] = 708
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'ADD_DATA_COLLECTION:\n  id: 64\n  fields:\n    collection: DATA_COLLECTION'"
+        "'DATASET_ADD:\n  id: 64\n  fields:\n    dataset: DATASET'"
     )
 
-    collection: Struct[DATA_COLLECTION] = Struct(DATA_COLLECTION)
+    dataset: Struct[DATASET] = Struct(DATASET)
 
 
 @pyrtma.message_def
-class MDF_REMOVE_DATA_COLLECTION(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_ADDED(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 65
-    type_name: ClassVar[str] = "REMOVE_DATA_COLLECTION"
-    type_hash: ClassVar[int] = 0x5430DCF2
+    type_name: ClassVar[str] = "DATASET_ADDED"
+    type_hash: ClassVar[int] = 0xCFA9917E
     type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'REMOVE_DATA_COLLECTION:\n  id: 65\n  fields:\n    collection_name: char[32]'"
+        "'DATASET_ADDED:\n  id: 65\n  fields:\n    name: char[DATASET_NAME_LEN]'"
     )
 
-    collection_name: String = String(32)
-
-
-@pyrtma.message_def
-class MDF_ADD_DATA_SET(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 66
-    type_name: ClassVar[str] = "ADD_DATA_SET"
-    type_hash: ClassVar[int] = 0xB368D74F
-    type_size: ClassVar[int] = 484
-    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'ADD_DATA_SET:\n  id: 66\n  fields:\n    collection_name: char[32]\n    data_set: DATA_SET'"
-    )
-
-    collection_name: String = String(32)
-    data_set: Struct[DATA_SET] = Struct(DATA_SET)
-
-
-@pyrtma.message_def
-class MDF_REMOVE_DATA_SET(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 67
-    type_name: ClassVar[str] = "REMOVE_DATA_SET"
-    type_hash: ClassVar[int] = 0x0BC39C04
-    type_size: ClassVar[int] = 64
-    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'REMOVE_DATA_SET:\n  id: 67\n  fields:\n    collection_name: char[32]\n    name: char[32]'"
-    )
-
-    collection_name: String = String(32)
     name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_COLLECTION_CONFIG_REQUEST(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_REMOVE(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 66
+    type_name: ClassVar[str] = "DATASET_REMOVE"
+    type_hash: ClassVar[int] = 0x13EEE304
+    type_size: ClassVar[int] = 32
+    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
+    type_def: ClassVar[str] = (
+        "'DATASET_REMOVE:\n  id: 66\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
+
+
+@pyrtma.message_def
+class MDF_DATASET_REMOVED(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 67
+    type_name: ClassVar[str] = "DATASET_REMOVED"
+    type_hash: ClassVar[int] = 0x7F388965
+    type_size: ClassVar[int] = 32
+    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
+    type_def: ClassVar[str] = (
+        "'DATASET_REMOVED:\n  id: 67\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
+
+
+@pyrtma.message_def
+class MDF_DATA_LOGGER_CONFIG_REQUEST(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 68
-    type_name: ClassVar[str] = "DATA_COLLECTION_CONFIG_REQUEST"
-    type_hash: ClassVar[int] = 0x1779D826
+    type_name: ClassVar[str] = "DATA_LOGGER_CONFIG_REQUEST"
+    type_hash: ClassVar[int] = 0x91598F0A
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_COLLECTION_CONFIG_REQUEST:\n  id: 68\n  fields: null'"
-    )
+    type_def: ClassVar[str] = "'DATA_LOGGER_CONFIG_REQUEST:\n  id: 68\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF_DATA_COLLECTION_CONFIG(MessageData, metaclass=MessageMeta):
+class MDF_DATA_LOGGER_CONFIG(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 69
-    type_name: ClassVar[str] = "DATA_COLLECTION_CONFIG"
-    type_hash: ClassVar[int] = 0xC08D3576
-    type_size: ClassVar[int] = 3132
+    type_name: ClassVar[str] = "DATA_LOGGER_CONFIG"
+    type_hash: ClassVar[int] = 0x085048FB
+    type_size: ClassVar[int] = 4252
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_COLLECTION_CONFIG:\n  id: 69\n  fields:\n    collection: DATA_COLLECTION'"
+        "'DATA_LOGGER_CONFIG:\n  id: 69\n  fields:\n    num_datasets: int32\n    datasets: DATASET[6]'"
     )
 
-    collection: Struct[DATA_COLLECTION] = Struct(DATA_COLLECTION)
+    num_datasets: Int32 = Int32()
+    datasets: StructArray[DATASET] = StructArray(DATASET, 6)
 
 
 @pyrtma.message_def
-class MDF_DATA_COLLECTION_STARTED(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_STARTED(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 70
-    type_name: ClassVar[str] = "DATA_COLLECTION_STARTED"
-    type_hash: ClassVar[int] = 0x415737CA
-    type_size: ClassVar[int] = 3004
+    type_name: ClassVar[str] = "DATASET_STARTED"
+    type_hash: ClassVar[int] = 0x5A6BC4E6
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_COLLECTION_STARTED:\n  id: 70\n  fields:\n    collection: DATA_COLLECTION_INFO'"
+        "'DATASET_STARTED:\n  id: 70\n  fields:\n    name: char[DATASET_NAME_LEN]'"
     )
 
-    collection: Struct[DATA_COLLECTION_INFO] = Struct(DATA_COLLECTION_INFO)
+    name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_COLLECTION_STOPPED(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_STOPPED(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 71
-    type_name: ClassVar[str] = "DATA_COLLECTION_STOPPED"
-    type_hash: ClassVar[int] = 0xDA1AF3AE
-    type_size: ClassVar[int] = 3004
+    type_name: ClassVar[str] = "DATASET_STOPPED"
+    type_hash: ClassVar[int] = 0x19F8847F
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_COLLECTION_STOPPED:\n  id: 71\n  fields:\n    collection: DATA_COLLECTION_INFO'"
+        "'DATASET_STOPPED:\n  id: 71\n  fields:\n    name: char[DATASET_NAME_LEN]'"
     )
 
-    collection: Struct[DATA_COLLECTION_INFO] = Struct(DATA_COLLECTION_INFO)
+    name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_COLLECTION_SAVED(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_SAVED(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 72
-    type_name: ClassVar[str] = "DATA_COLLECTION_SAVED"
-    type_hash: ClassVar[int] = 0x74528518
-    type_size: ClassVar[int] = 0
+    type_name: ClassVar[str] = "DATASET_SAVED"
+    type_hash: ClassVar[int] = 0x964F725E
+    type_size: ClassVar[int] = 1056
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_COLLECTION_SAVED:\n  id: 72\n  fields: null'"
+    type_def: ClassVar[str] = (
+        "'DATASET_SAVED:\n  id: 72\n  fields:\n    name: char[DATASET_NAME_LEN]\n    filepath: char[1024]'"
+    )
+
+    name: String = String(32)
+    filepath: String = String(1024)
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_START(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_START(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 73
-    type_name: ClassVar[str] = "DATA_LOGGER_START"
-    type_hash: ClassVar[int] = 0x982D6942
-    type_size: ClassVar[int] = 0
+    type_name: ClassVar[str] = "DATASET_START"
+    type_hash: ClassVar[int] = 0x019BBFB4
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_LOGGER_START:\n  id: 73\n  fields: null'"
+    type_def: ClassVar[str] = (
+        "'DATASET_START:\n  id: 73\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_STOP(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_STOP(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 74
-    type_name: ClassVar[str] = "DATA_LOGGER_STOP"
-    type_hash: ClassVar[int] = 0xA5B66B24
-    type_size: ClassVar[int] = 0
+    type_name: ClassVar[str] = "DATASET_STOP"
+    type_hash: ClassVar[int] = 0xFC4A83B1
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_LOGGER_STOP:\n  id: 74\n  fields: null'"
+    type_def: ClassVar[str] = (
+        "'DATASET_STOP:\n  id: 74\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_PAUSE(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_PAUSE(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 75
-    type_name: ClassVar[str] = "DATA_LOGGER_PAUSE"
-    type_hash: ClassVar[int] = 0x9A556835
-    type_size: ClassVar[int] = 0
+    type_name: ClassVar[str] = "DATASET_PAUSE"
+    type_hash: ClassVar[int] = 0x0DCFA7B0
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_LOGGER_PAUSE:\n  id: 75\n  fields: null'"
+    type_def: ClassVar[str] = (
+        "'DATASET_PAUSE:\n  id: 75\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_RESUME(MessageData, metaclass=MessageMeta):
+class MDF_DATASET_RESUME(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 76
-    type_name: ClassVar[str] = "DATA_LOGGER_RESUME"
-    type_hash: ClassVar[int] = 0x18811751
-    type_size: ClassVar[int] = 0
+    type_name: ClassVar[str] = "DATASET_RESUME"
+    type_hash: ClassVar[int] = 0xF4E01F27
+    type_size: ClassVar[int] = 32
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = "'DATA_LOGGER_RESUME:\n  id: 76\n  fields: null'"
+    type_def: ClassVar[str] = (
+        "'DATASET_RESUME:\n  id: 76\n  fields:\n    name: char[DATASET_NAME_LEN]'"
+    )
+
+    name: String = String(32)
 
 
 @pyrtma.message_def
@@ -949,65 +925,27 @@ class MDF_DATA_LOGGER_RESET(MessageData, metaclass=MessageMeta):
 class MDF_DATA_LOGGER_ERROR(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 78
     type_name: ClassVar[str] = "DATA_LOGGER_ERROR"
-    type_hash: ClassVar[int] = 0xAD3B4A67
-    type_size: ClassVar[int] = 512
+    type_hash: ClassVar[int] = 0x630C7107
+    type_size: ClassVar[int] = 608
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_LOGGER_ERROR:\n  id: 78\n  fields:\n    msg: char[512]'"
+        "'DATA_LOGGER_ERROR:\n  id: 78\n  fields:\n    dataset_name: char[DATASET_NAME_LEN]\n    exc_type: char[64]\n    msg: char[512]'"
     )
 
+    dataset_name: String = String(32)
+    exc_type: String = String(64)
     msg: String = String(512)
 
 
 @pyrtma.message_def
-class MDF_DATA_LOGGER_METADATA_UPDATE(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 79
-    type_name: ClassVar[str] = "DATA_LOGGER_METADATA_UPDATE"
-    type_hash: ClassVar[int] = 0x3C673526
-    type_size: ClassVar[int] = 512
-    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_LOGGER_METADATA_UPDATE:\n  id: 79\n  fields:\n    json: char[512]'"
-    )
-
-    json: String = String(512)
-
-
-@pyrtma.message_def
-class MDF_DATA_LOGGER_METADATA_REQUEST(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 87
-    type_name: ClassVar[str] = "DATA_LOGGER_METADATA_REQUEST"
-    type_hash: ClassVar[int] = 0x2D068003
-    type_size: ClassVar[int] = 0
-    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_LOGGER_METADATA_REQUEST:\n  id: 87\n  fields: null'"
-    )
-
-
-@pyrtma.message_def
-class MDF_DATA_LOGGER_METADATA(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 88
-    type_name: ClassVar[str] = "DATA_LOGGER_METADATA"
-    type_hash: ClassVar[int] = 0x3852789B
-    type_size: ClassVar[int] = 1024
-    type_source: ClassVar[str] = "core_defs/data_logger.yaml"
-    type_def: ClassVar[str] = (
-        "'DATA_LOGGER_METADATA:\n  id: 88\n  fields:\n    json: char[1024]'"
-    )
-
-    json: String = String(1024)
-
-
-@pyrtma.message_def
 class MDF_DATA_LOG_TEST_2048(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 89
+    type_id: ClassVar[int] = 79
     type_name: ClassVar[str] = "DATA_LOG_TEST_2048"
-    type_hash: ClassVar[int] = 0x651D3C98
+    type_hash: ClassVar[int] = 0x6F24B1D1
     type_size: ClassVar[int] = 2048
     type_source: ClassVar[str] = "core_defs/data_logger.yaml"
     type_def: ClassVar[str] = (
-        "'DATA_LOG_TEST_2048:\n  id: 89\n  fields:\n    raw: char[2048]'"
+        "'DATA_LOG_TEST_2048:\n  id: 79\n  fields:\n    raw: char[2048]'"
     )
 
     raw: String = String(2048)
@@ -1110,6 +1048,16 @@ class MDF_DUMP_MESSAGE_LOG(MessageData, metaclass=MessageMeta):
 
 
 @pyrtma.message_def
+class MDF_DEBUG_TEXT(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 91
+    type_name: ClassVar[str] = "DEBUG_TEXT"
+    type_hash: ClassVar[int] = 0xEC89E79E
+    type_size: ClassVar[int] = 0
+    type_source: ClassVar[str] = "core_defs/quick_logger.yaml"
+    type_def: ClassVar[str] = "'DEBUG_TEXT:\n  id: 91\n  fields: null'"
+
+
+@pyrtma.message_def
 class MDF_LM_READY(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 96
     type_name: ClassVar[str] = "LM_READY"
@@ -1185,6 +1133,22 @@ class MDF_FAIL_SUBSCRIBE(MessageData, metaclass=MessageMeta):
 
 
 @pyrtma.message_def
+class MDF_SUBSCRIPTION_OPTION(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 7
+    type_name: ClassVar[str] = "SUBSCRIPTION_OPTION"
+    type_hash: ClassVar[int] = 0xFF6CEDFC
+    type_size: ClassVar[int] = 12
+    type_source: ClassVar[str] = "core_defs/core_defs.yaml"
+    type_def: ClassVar[str] = (
+        "'SUBSCRIPTION_OPTION:\n  id: 7\n  fields:\n    msg_type: MSG_TYPE\n    option: int32\n    value: int32'"
+    )
+
+    msg_type: Int32 = Int32()
+    option: Int32 = Int32()
+    value: Int32 = Int32()
+
+
+@pyrtma.message_def
 class MDF_FAILED_MESSAGE(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 8
     type_name: ClassVar[str] = "FAILED_MESSAGE"
@@ -1199,6 +1163,23 @@ class MDF_FAILED_MESSAGE(MessageData, metaclass=MessageMeta):
     reserved: IntArray[Int16] = IntArray(Int16, 3)
     time_of_failure: Double = Double()
     msg_header: Struct[RTMA_MSG_HEADER] = Struct(RTMA_MSG_HEADER)
+
+
+@pyrtma.message_def
+class MDF_MESSAGE_TIMING(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 9
+    type_name: ClassVar[str] = "MESSAGE_TIMING"
+    type_hash: ClassVar[int] = 0xE001C54C
+    type_size: ClassVar[int] = 16
+    type_source: ClassVar[str] = "core_defs/core_defs.yaml"
+    type_def: ClassVar[str] = (
+        "'MESSAGE_TIMING:\n  id: 9\n  fields:\n    msg_type: MSG_TYPE\n    src_id: MODULE_ID\n    reserved: int16\n    send_time: double'"
+    )
+
+    msg_type: Int32 = Int32()
+    src_id: Int16 = Int16()
+    reserved: Int16 = Int16()
+    send_time: Double = Double()
 
 
 @pyrtma.message_def
@@ -1255,16 +1236,6 @@ class MDF_UNSUBSCRIBE(MessageData, metaclass=MessageMeta):
 
 
 @pyrtma.message_def
-class MDF_SHUTDOWN_RTMA(MessageData, metaclass=MessageMeta):
-    type_id: ClassVar[int] = 17
-    type_name: ClassVar[str] = "SHUTDOWN_RTMA"
-    type_hash: ClassVar[int] = 0x0912BFA5
-    type_size: ClassVar[int] = 0
-    type_source: ClassVar[str] = "core_defs/core_defs.yaml"
-    type_def: ClassVar[str] = "'SHUTDOWN_RTMA:\n  id: 17\n  fields: null'"
-
-
-@pyrtma.message_def
 class MDF_MODULE_READY(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 26
     type_name: ClassVar[str] = "MODULE_READY"
@@ -1277,82 +1248,72 @@ class MDF_MODULE_READY(MessageData, metaclass=MessageMeta):
 
 
 @pyrtma.message_def
-class MDF_MESSAGE_TRAFFIC(MessageData, metaclass=MessageMeta):
+class MDF_PING(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 29
+    type_name: ClassVar[str] = "PING"
+    type_hash: ClassVar[int] = 0x43014E6F
+    type_size: ClassVar[int] = 4
+    type_source: ClassVar[str] = "core_defs/core_defs.yaml"
+    type_def: ClassVar[str] = "'PING:\n  id: 29\n  fields:\n    uid: int32'"
+
+    uid: Int32 = Int32()
+
+
+@pyrtma.message_def
+class MDF_PONG(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 30
-    type_name: ClassVar[str] = "MESSAGE_TRAFFIC"
-    type_hash: ClassVar[int] = 0xA5D05FF5
-    type_size: ClassVar[int] = 408
+    type_name: ClassVar[str] = "PONG"
+    type_hash: ClassVar[int] = 0xDAC22B06
+    type_size: ClassVar[int] = 4
     type_source: ClassVar[str] = "core_defs/core_defs.yaml"
-    type_def: ClassVar[str] = (
-        "'MESSAGE_TRAFFIC:\n  id: 30\n  fields:\n    seqno: uint32\n    sub_seqno: uint32\n    start_timestamp: double\n    end_timestamp: double\n    msg_type: MSG_TYPE[MESSAGE_TRAFFIC_SIZE]\n    msg_count: uint16[MESSAGE_TRAFFIC_SIZE]'"
-    )
+    type_def: ClassVar[str] = "'PONG:\n  id: 30\n  fields:\n    uid: int32'"
 
-    seqno: Uint32 = Uint32()
-    sub_seqno: Uint32 = Uint32()
-    start_timestamp: Double = Double()
-    end_timestamp: Double = Double()
-    msg_type: IntArray[Int32] = IntArray(Int32, 64)
-    msg_count: IntArray[Uint16] = IntArray(Uint16, 64)
+    uid: Int32 = Int32()
 
 
 @pyrtma.message_def
-class MDF_ACTIVE_CLIENTS(MessageData, metaclass=MessageMeta):
+class MDF_INTRODUCE(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 31
-    type_name: ClassVar[str] = "ACTIVE_CLIENTS"
-    type_hash: ClassVar[int] = 0xFC42B3AF
-    type_size: ClassVar[int] = 1552
+    type_name: ClassVar[str] = "INTRODUCE"
+    type_hash: ClassVar[int] = 0x6F4769E8
+    type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "core_defs/core_defs.yaml"
-    type_def: ClassVar[str] = (
-        "'ACTIVE_CLIENTS:\n  id: 31\n  fields:\n    timestamp: double\n    num_clients: int16\n    padding: int16\n    reserved: int32\n    client_mod_id: MODULE_ID[MAX_ACTIVE_CLIENTS]\n    client_pid: int32[MAX_ACTIVE_CLIENTS]'"
-    )
-
-    timestamp: Double = Double()
-    num_clients: Int16 = Int16()
-    padding: Int16 = Int16()
-    reserved: Int32 = Int32()
-    client_mod_id: IntArray[Int16] = IntArray(Int16, 256)
-    client_pid: IntArray[Int32] = IntArray(Int32, 256)
+    type_def: ClassVar[str] = "'INTRODUCE:\n  id: 31\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF_CLIENT_INFO(MessageData, metaclass=MessageMeta):
+class MDF_HELLO(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 32
-    type_name: ClassVar[str] = "CLIENT_INFO"
-    type_hash: ClassVar[int] = 0x401076E7
-    type_size: ClassVar[int] = 80
+    type_name: ClassVar[str] = "HELLO"
+    type_hash: ClassVar[int] = 0x1BDE1402
+    type_size: ClassVar[int] = 44
     type_source: ClassVar[str] = "core_defs/core_defs.yaml"
     type_def: ClassVar[str] = (
-        "'CLIENT_INFO:\n  id: 32\n  fields:\n    addr: char[32]\n    uid: int32\n    pid: int32\n    mod_id: MODULE_ID\n    is_logger: int16\n    is_unique: int16\n    port: uint16\n    name: char[MAX_NAME_LEN]'"
+        "'HELLO:\n  id: 32\n  fields:\n    uid: int32\n    pid: int32\n    mod_id: MODULE_ID\n    reserved: int16\n    name: char[MAX_NAME_LEN]'"
     )
 
-    addr: String = String(32)
     uid: Int32 = Int32()
     pid: Int32 = Int32()
     mod_id: Int16 = Int16()
-    is_logger: Int16 = Int16()
-    is_unique: Int16 = Int16()
-    port: Uint16 = Uint16()
+    reserved: Int16 = Int16()
     name: String = String(32)
 
 
 @pyrtma.message_def
-class MDF_CLIENT_CLOSED(MessageData, metaclass=MessageMeta):
+class MDF_GOODBYE(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 33
-    type_name: ClassVar[str] = "CLIENT_CLOSED"
-    type_hash: ClassVar[int] = 0x087F4249
-    type_size: ClassVar[int] = 80
+    type_name: ClassVar[str] = "GOODBYE"
+    type_hash: ClassVar[int] = 0x9998622C
+    type_size: ClassVar[int] = 44
     type_source: ClassVar[str] = "core_defs/core_defs.yaml"
     type_def: ClassVar[str] = (
-        "'CLIENT_CLOSED:\n  id: 33\n  fields:\n    addr: char[32]\n    uid: int32\n    pid: int32\n    mod_id: MODULE_ID\n    is_logger: int16\n    is_unique: int16\n    port: uint16\n    name: char[MAX_NAME_LEN]'"
+        "'GOODBYE:\n  id: 33\n  fields:\n    uid: int32\n    pid: int32\n    mod_id: MODULE_ID\n    reserved: int16\n    name: char[MAX_NAME_LEN]'"
     )
 
-    addr: String = String(32)
     uid: Int32 = Int32()
     pid: Int32 = Int32()
     mod_id: Int16 = Int16()
-    is_logger: Int16 = Int16()
-    is_unique: Int16 = Int16()
-    port: Uint16 = Uint16()
+    reserved: Int16 = Int16()
     name: String = String(32)
 
 
@@ -2217,10 +2178,10 @@ class MDF_GRIP_COMMAND(MessageData, metaclass=MessageMeta):
     )
 
     header: Struct[MSG_HEADER] = Struct(MSG_HEADER)
-    grip_pos: Double = Double()
-    velocity: Double = Double()
-    force: Double = Double()
-    impedance: Double = Double()
+    grip_pos: FloatArray[Double] = FloatArray(Double, 1)
+    velocity: FloatArray[Double] = FloatArray(Double, 1)
+    force: FloatArray[Double] = FloatArray(Double, 1)
+    impedance: FloatArray[Double] = FloatArray(Double, 1)
     controlMask: IntArray[Int16] = IntArray(Int16, 4)
     src: Int32 = Int32()
     reserved: Int32 = Int32()
@@ -2238,10 +2199,10 @@ class MDF_GRIP_FINISHED_COMMAND(MessageData, metaclass=MessageMeta):
     )
 
     header: Struct[MSG_HEADER] = Struct(MSG_HEADER)
-    grip_pos: Double = Double()
-    velocity: Double = Double()
-    force: Double = Double()
-    impedance: Double = Double()
+    grip_pos: FloatArray[Double] = FloatArray(Double, 1)
+    velocity: FloatArray[Double] = FloatArray(Double, 1)
+    force: FloatArray[Double] = FloatArray(Double, 1)
+    impedance: FloatArray[Double] = FloatArray(Double, 1)
     controlMask: IntArray[Int16] = IntArray(Int16, 4)
     effector: String = String(64)
 
@@ -2258,8 +2219,8 @@ class MDF_GRIPPER_FEEDBACK(MessageData, metaclass=MessageMeta):
     )
 
     header: Struct[MSG_HEADER] = Struct(MSG_HEADER)
-    grip_pos: Double = Double()
-    velocity: Double = Double()
+    grip_pos: FloatArray[Double] = FloatArray(Double, 1)
+    velocity: FloatArray[Double] = FloatArray(Double, 1)
     force: FloatArray[Double] = FloatArray(Double, 5)
     effector: String = String(64)
 
@@ -2296,10 +2257,10 @@ class MDF_MUJOCO_CMD(MessageData, metaclass=MessageMeta):
     )
 
     header: Struct[MSG_HEADER] = Struct(MSG_HEADER)
-    ref_pos: Double = Double()
-    ref_vel: Double = Double()
-    gain_pos: Double = Double()
-    gain_vel: Double = Double()
+    ref_pos: FloatArray[Double] = FloatArray(Double, 1)
+    ref_vel: FloatArray[Double] = FloatArray(Double, 1)
+    gain_pos: FloatArray[Double] = FloatArray(Double, 1)
+    gain_vel: FloatArray[Double] = FloatArray(Double, 1)
     ref_pos_enabled: Int16 = Int16()
     ref_vel_enabled: Int16 = Int16()
     gain_pos_enabled: Int16 = Int16()
@@ -4780,203 +4741,203 @@ class MDF_UC_MECH_STIM_CONFIGURE(MessageData, metaclass=MessageMeta):
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004300(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004300(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4300
-    type_name: ClassVar[str] = "_RESERVED_004300"
-    type_hash: ClassVar[int] = 0xCE2B74C1
+    type_name: ClassVar[str] = "RESERVED_004300"
+    type_hash: ClassVar[int] = 0x2FB5544D
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004300:\n  id: 4300\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004300:\n  id: 4300\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004360(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004360(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4360
-    type_name: ClassVar[str] = "_RESERVED_004360"
-    type_hash: ClassVar[int] = 0xA4288CBE
+    type_name: ClassVar[str] = "RESERVED_004360"
+    type_hash: ClassVar[int] = 0x323B7BC7
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004360:\n  id: 4360\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004360:\n  id: 4360\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004361(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004361(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4361
-    type_name: ClassVar[str] = "_RESERVED_004361"
-    type_hash: ClassVar[int] = 0xCA551012
+    type_name: ClassVar[str] = "RESERVED_004361"
+    type_hash: ClassVar[int] = 0xCEA6EA95
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004361:\n  id: 4361\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004361:\n  id: 4361\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004362(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004362(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4362
-    type_name: ClassVar[str] = "_RESERVED_004362"
-    type_hash: ClassVar[int] = 0x1FB4158D
+    type_name: ClassVar[str] = "RESERVED_004362"
+    type_hash: ClassVar[int] = 0x02A36526
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004362:\n  id: 4362\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004362:\n  id: 4362\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004363(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004363(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4363
-    type_name: ClassVar[str] = "_RESERVED_004363"
-    type_hash: ClassVar[int] = 0x6A1BE0D0
+    type_name: ClassVar[str] = "RESERVED_004363"
+    type_hash: ClassVar[int] = 0xECC4954A
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004363:\n  id: 4363\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004363:\n  id: 4363\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004364(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004364(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4364
-    type_name: ClassVar[str] = "_RESERVED_004364"
-    type_hash: ClassVar[int] = 0xC0CA6671
+    type_name: ClassVar[str] = "RESERVED_004364"
+    type_hash: ClassVar[int] = 0x7910FD63
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004364:\n  id: 4364\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004364:\n  id: 4364\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004365(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004365(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4365
-    type_name: ClassVar[str] = "_RESERVED_004365"
-    type_hash: ClassVar[int] = 0x862615BB
+    type_name: ClassVar[str] = "RESERVED_004365"
+    type_hash: ClassVar[int] = 0x935F393E
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004365:\n  id: 4365\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004365:\n  id: 4365\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004366(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004366(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4366
-    type_name: ClassVar[str] = "_RESERVED_004366"
-    type_hash: ClassVar[int] = 0xA703073C
+    type_name: ClassVar[str] = "RESERVED_004366"
+    type_hash: ClassVar[int] = 0xD519DA6B
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004366:\n  id: 4366\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004366:\n  id: 4366\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004367(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004367(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4367
-    type_name: ClassVar[str] = "_RESERVED_004367"
-    type_hash: ClassVar[int] = 0x442ABB7F
+    type_name: ClassVar[str] = "RESERVED_004367"
+    type_hash: ClassVar[int] = 0x08E1386F
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004367:\n  id: 4367\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004367:\n  id: 4367\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004368(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004368(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4368
-    type_name: ClassVar[str] = "_RESERVED_004368"
-    type_hash: ClassVar[int] = 0x8CD5099C
+    type_name: ClassVar[str] = "RESERVED_004368"
+    type_hash: ClassVar[int] = 0x99F64A37
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004368:\n  id: 4368\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004368:\n  id: 4368\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004369(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004369(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4369
-    type_name: ClassVar[str] = "_RESERVED_004369"
-    type_hash: ClassVar[int] = 0xC91AF2B3
+    type_name: ClassVar[str] = "RESERVED_004369"
+    type_hash: ClassVar[int] = 0xF4B70AB4
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004369:\n  id: 4369\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004369:\n  id: 4369\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004370(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004370(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4370
-    type_name: ClassVar[str] = "_RESERVED_004370"
-    type_hash: ClassVar[int] = 0x1A8E353C
+    type_name: ClassVar[str] = "RESERVED_004370"
+    type_hash: ClassVar[int] = 0xE46A53AF
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004370:\n  id: 4370\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004370:\n  id: 4370\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004371(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004371(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4371
-    type_name: ClassVar[str] = "_RESERVED_004371"
-    type_hash: ClassVar[int] = 0x09248DD4
+    type_name: ClassVar[str] = "RESERVED_004371"
+    type_hash: ClassVar[int] = 0x177FBC2F
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004371:\n  id: 4371\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004371:\n  id: 4371\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004372(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004372(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4372
-    type_name: ClassVar[str] = "_RESERVED_004372"
-    type_hash: ClassVar[int] = 0x9880225F
+    type_name: ClassVar[str] = "RESERVED_004372"
+    type_hash: ClassVar[int] = 0x01736936
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004372:\n  id: 4372\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004372:\n  id: 4372\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004373(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004373(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4373
-    type_name: ClassVar[str] = "_RESERVED_004373"
-    type_hash: ClassVar[int] = 0xA9E362D1
+    type_name: ClassVar[str] = "RESERVED_004373"
+    type_hash: ClassVar[int] = 0x72CF2D3F
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004373:\n  id: 4373\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004373:\n  id: 4373\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004374(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004374(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4374
-    type_name: ClassVar[str] = "_RESERVED_004374"
-    type_hash: ClassVar[int] = 0xF76E2A3C
+    type_name: ClassVar[str] = "RESERVED_004374"
+    type_hash: ClassVar[int] = 0xEC101B55
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004374:\n  id: 4374\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004374:\n  id: 4374\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004375(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004375(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4375
-    type_name: ClassVar[str] = "_RESERVED_004375"
-    type_hash: ClassVar[int] = 0xB01FBB58
+    type_name: ClassVar[str] = "RESERVED_004375"
+    type_hash: ClassVar[int] = 0xB41F9FC1
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004375:\n  id: 4375\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004375:\n  id: 4375\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004376(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004376(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4376
-    type_name: ClassVar[str] = "_RESERVED_004376"
-    type_hash: ClassVar[int] = 0xC4D255ED
+    type_name: ClassVar[str] = "RESERVED_004376"
+    type_hash: ClassVar[int] = 0xEBC7B874
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004376:\n  id: 4376\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004376:\n  id: 4376\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004377(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004377(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4377
-    type_name: ClassVar[str] = "_RESERVED_004377"
-    type_hash: ClassVar[int] = 0x5DB1542A
+    type_name: ClassVar[str] = "RESERVED_004377"
+    type_hash: ClassVar[int] = 0x4DCCD750
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004377:\n  id: 4377\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004377:\n  id: 4377\n  fields: null'"
 
 
 @pyrtma.message_def
-class MDF__RESERVED_004378(MessageData, metaclass=MessageMeta):
+class MDF_RESERVED_004378(MessageData, metaclass=MessageMeta):
     type_id: ClassVar[int] = 4378
-    type_name: ClassVar[str] = "_RESERVED_004378"
-    type_hash: ClassVar[int] = 0xD323D536
+    type_name: ClassVar[str] = "RESERVED_004378"
+    type_hash: ClassVar[int] = 0x75FB28B0
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
-    type_def: ClassVar[str] = "'_RESERVED_004378:\n  id: 4378\n  fields: null'"
+    type_def: ClassVar[str] = "'RESERVED_004378:\n  id: 4378\n  fields: null'"
 
 
 @pyrtma.message_def
@@ -4996,24 +4957,24 @@ class MDF_VALIDATOR_A(MessageData, metaclass=MessageMeta):
     int32: Int32 = Int32()
     int64: Int64 = Int64()
     uint8: Uint8 = Uint8()
-    padding_0_: Char = Char()
+    padding_0_: ByteArray = ByteArray(1)
     uint16: Uint16 = Uint16()
     uint32: Uint32 = Uint32()
     uint64: Uint64 = Uint64()
     int8_arr: IntArray[Int8] = IntArray(Int8, 4)
     int16_arr: IntArray[Int16] = IntArray(Int16, 4)
     int32_arr: IntArray[Int32] = IntArray(Int32, 4)
-    padding_1_: String = String(4)
+    padding_1_: ByteArray = ByteArray(4)
     int64_arr: IntArray[Int64] = IntArray(Int64, 4)
     uint8_arr: IntArray[Uint8] = IntArray(Uint8, 4)
     uint16_arr: IntArray[Uint16] = IntArray(Uint16, 4)
     uint32_arr: IntArray[Uint32] = IntArray(Uint32, 4)
-    padding_2_: String = String(4)
+    padding_2_: ByteArray = ByteArray(4)
     uint64_arr: IntArray[Uint64] = IntArray(Uint64, 4)
     byte_: Byte = Byte()
     byte_arr: ByteArray = ByteArray(4)
     string: String = String(4)
-    padding_3_: String = String(3)
+    padding_3_: ByteArray = ByteArray(3)
     float_: Float = Float()
     double_: Double = Double()
     float_arr: FloatArray[Float] = FloatArray(Float, 4)
@@ -5039,7 +5000,7 @@ class MDF_VALIDATOR_B(MessageData, metaclass=MessageMeta):
     int32: Int32 = Int32()
     int64: Int64 = Int64()
     uint8: Uint8 = Uint8()
-    padding_0_: Char = Char()
+    padding_0_: ByteArray = ByteArray(1)
     uint16: Uint16 = Uint16()
     uint32: Uint32 = Uint32()
     uint64: Uint64 = Uint64()
@@ -5054,7 +5015,7 @@ class MDF_VALIDATOR_B(MessageData, metaclass=MessageMeta):
     byte_: Byte = Byte()
     byte_arr: ByteArray = ByteArray(8)
     string: String = String(8)
-    padding_1_: String = String(3)
+    padding_1_: ByteArray = ByteArray(3)
     float_: Float = Float()
     double_: Double = Double()
     float_arr: FloatArray[Float] = FloatArray(Float, 8)
@@ -5199,6 +5160,39 @@ class MDF_PUBLISHER_DONE(MessageData, metaclass=MessageMeta):
     type_size: ClassVar[int] = 0
     type_source: ClassVar[str] = "test_defs.yaml"
     type_def: ClassVar[str] = "'PUBLISHER_DONE:\n  id: 5012\n  fields: null'"
+
+
+@pyrtma.message_def
+class MDF_ARRAY_OF_ONE(MessageData, metaclass=MessageMeta):
+    type_id: ClassVar[int] = 5013
+    type_name: ClassVar[str] = "ARRAY_OF_ONE"
+    type_hash: ClassVar[int] = 0x1BE0A085
+    type_size: ClassVar[int] = 288
+    type_source: ClassVar[str] = "test_defs.yaml"
+    type_def: ClassVar[str] = (
+        "'ARRAY_OF_ONE:\n  id: 5013\n  fields:\n    u8: uint8[1]\n    u16: uint16[1]\n    u32: uint32[1]\n    u64: uint64[1]\n    i8: int8[1]\n    i16: int16[1]\n    i32: int32[1]\n    i64: int64[1]\n    f32: float[1]\n    f64: double[1]\n    c: char\n    c1: char[1]\n    s2: char[2]\n    b: byte\n    ba: byte[1]\n    sa: VALIDATOR_STRUCT[1]'"
+    )
+
+    u8: IntArray[Uint8] = IntArray(Uint8, 1)
+    padding_0_: ByteArray = ByteArray(1)
+    u16: IntArray[Uint16] = IntArray(Uint16, 1)
+    u32: IntArray[Uint32] = IntArray(Uint32, 1)
+    u64: IntArray[Uint64] = IntArray(Uint64, 1)
+    i8: IntArray[Int8] = IntArray(Int8, 1)
+    padding_1_: ByteArray = ByteArray(1)
+    i16: IntArray[Int16] = IntArray(Int16, 1)
+    i32: IntArray[Int32] = IntArray(Int32, 1)
+    i64: IntArray[Int64] = IntArray(Int64, 1)
+    f32: FloatArray[Float] = FloatArray(Float, 1)
+    padding_2_: ByteArray = ByteArray(4)
+    f64: FloatArray[Double] = FloatArray(Double, 1)
+    c: Char = Char()
+    c1: Char = Char()
+    s2: String = String(2)
+    b: Byte = Byte()
+    ba: ByteArray = ByteArray(1)
+    padding_3_: ByteArray = ByteArray(2)
+    sa: StructArray[VALIDATOR_STRUCT] = StructArray(VALIDATOR_STRUCT, 1)
 
 
 # User Context
